added infrastructure to rename slashed and frozen nonterminals when necessary
added CKY parsing methods

BUGFIX: CNF conversion now correctly ensures that the start symbol can only
appear on the LHS of the grammar rules; added the `separate_start` method to
accomplish this and it is called in the CNF conversion pipeline.

API change: `unfold` now takes an integer rule id instead of a reference to the
Rule.

added default values for `Ps` in `speculation` and `lc_generalized`.

update tutorial notebook and test cases to reflect these changes.